pyproject.toml
[project]
name = "python-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2026.4.22",
    "google-genai>=1.73.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.11.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
